refactor: Extract duplicated pattern matching to shared utility
Created internal/core/pattern.go with unified MatchesPattern function.
This consolidates the duplicate implementations from:
- internal/core/state.go (matchesPattern - private)
- internal/commands/archive.go (MatchesPattern - public)

The unified version adopts the more robust behavior from archive.go:
- Returns false for empty patternParts (no pattern = no match)
- Handles "*" wildcard as special case (matches everything)

Updated tests to use the shared function from core package.

Fixes #68
